pyproject.toml
[tool.poetry]
name = "tiro"
version = "0.1.1"
description = ""
authors = ["Zhu Zhaomeng <zhaomeng.zhu@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = "^1.9.0"
Faker = "^13.3.4"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
typer = "^0.4.1"
PyYAML = "^6.0"
httpx = "^0.22.0"
rich = "^12.2.0"
nats-py = "^2.1.0"
karez = {path = "deps/karez-0.2.16-py3-none-any.whl"}
jsonschema = "^4.4.0"

[tool.poetry.dev-dependencies]
utinni = {path = "deps/utinni-0.2.1-py3-none-any.whl"}
jupyterlab = "^3.3.4"
influxdb-client = "^1.28.0"
python-arango = "^7.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tiro = "tiro.cli:app"
